Move the new functions to inventory, where they belong
