+++ components/generatedPeacockRequireTable.ts
@@ -39,7 +39,6 @@ import * as profileHandler from "./profileHandler"
 import * as scoreHandler from "./scoreHandler"
 import * as sessionSerialization from "./sessionSerialization"
 import * as tools from "./tools"
-import * as unlockables from "./unlockables"
 import * as utils from "./utils"
 import * as webFeatures from "./webFeatures"
 import * as legacyContractHandler from "./2016/legacyContractHandler"
@@ -138,7 +137,6 @@ export default {
         ...sessionSerialization,
     },
     "@peacockproject/core/tools": { __esModule: true, ...tools },
-    "@peacockproject/core/unlockables": { __esModule: true, ...unlockables },
     "@peacockproject/core/utils": { __esModule: true, ...utils },
     "@peacockproject/core/webFeatures": { __esModule: true, ...webFeatures },
     "@peacockproject/core/2016/legacyContractHandler": {
