+++ components/menus/planning.ts
@@ -41,14 +41,13 @@ import {
 } from "../utils"
 
 import type { Response } from "express"
-import { createInventory } from "../inventory"
+import { createInventory, getUnlockableById } from "../inventory"
 import { createSniperLoadouts } from "./sniper"
 import { getFlag } from "../flags"
 import { loadouts } from "../loadouts"
 import { resolveProfiles } from "../profileHandler"
 import { PlanningQuery } from "../types/gameSchemas"
 import { userAuths } from "../officialServerAuth"
-import { getUnlockableById } from "../unlockables"
 
 export async function planningView(
     req: RequestWithJwt<PlanningQuery>,
